# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Livia API — multi-tenant appointment operating system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/wellness/reports:
    get:
      operationId: getWellnessReports
      tags: [wellness]
      summary: Wellness reports bundle (room heatmap, packages, stress score)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
      responses:
        "200":
          description: Reports bundle
        "401":
          $ref: "#/components/responses/Unauthorized"
  /businesses/{businessId}/wellness/eod-close:
    get:
      operationId: getWellnessEodClose
      tags: [wellness]
      summary: End-of-day close narrative for wellness operators
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
      responses:
        "200":
          description: EOD close lines and stats
        "401":
          $ref: "#/components/responses/Unauthorized"
  /businesses/{businessId}/wellness/audit-diary:
    get:
      operationId: getWellnessAuditDiary
      tags: [wellness]
      summary: Wellness audit diary (bookings, Liv memory, audit events)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
      responses:
        "200":
          description: Diary lines
        "401":
          $ref: "#/components/responses/Unauthorized"
  /businesses/{businessId}/wellness/guest-vault:
    get:
      operationId: getWellnessGuestVault
      tags: [wellness]
      summary: Guest vault profile by phone
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
        - name: phone
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Vault profile
        "400":
          $ref: "#/components/responses/BadRequest"
  /businesses/{businessId}/wellness/redeem-code:
    post:
      operationId: lookupWellnessRedeemCode
      tags: [wellness]
      summary: Look up gift or package redemption code
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: Ledger lookup
        "404":
          $ref: "#/components/responses/NotFound"
  /businesses/{businessId}/wellness/duty-solver:
    post:
      operationId: wellnessDutySolver
      tags: [wellness]
      summary: Find free therapists in a room at an hour
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                resourceName: { type: string }
                hour: { type: number }
      responses:
        "200":
          description: Solver matches
  /businesses/{businessId}/wellness/terminal/checkout:
    post:
      operationId: wellnessTerminalCheckout
      tags: [wellness]
      summary: Stripe Terminal checkout hook for desk payment
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/BusinessId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId]
              properties:
                bookingId: { type: string }
      responses:
        "200":
          description: Checkout initiated
...
  /public/b/{slug}/chat:
    post:
      operationId: sendPublicChatMessage
      tags: [chat]
      summary: Send a message to the public AI booking assistant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicChatBody"
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalOpsSecret:
      type: apiKey
      in: header
      name: X-Internal-Ops-Secret
      description: Livia Inc operator secret (INTERNAL_OPS_SECRET or INTERNAL_CRON_SECRET fallback)
  parameters:
    BusinessId:
      name: businessId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PublicChatBody:
      type: object
      properties:
        conversationId:
          type: string
        message:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
      required: [message]
    PublicChatResponse:
      type: object
      properties:
        conversationId:
          type: string
        reply:
          type: string
        bookingId:
          type: ["string", "null"]
        status:
          type: string
          enum: [OPEN, HANDED_OFF, CLOSED]
      required: [conversationId, reply, status]
...